fix: first rev fixes2
