+++ crate/findex/src/ovec.rs
@@ -229,6 +229,7 @@ mod tests {
         let seed = Secret::random(&mut rng);
         let memory = InMemory::<Address<ADDRESS_LENGTH>, [u8; WORD_LENGTH]>::default();
         let address = Address::random(&mut rng);
+        let obf = MemoryEncryptionLayer::new(&seed, memory.clone());
         let v = IVec::<WORD_LENGTH, _>::new(address, obf);
         test_vector_sequential(&v).await;
         memory.clear();
